refactored chain page into tabs & sections
