+++ chain-ui/components/ChainUiComponents__StatusIcons.tsx
@@ -1,0 +1,7 @@
+import { FiAlertCircle, FiCheckCircle, FiXCircle } from "react-icons/fi";
+
+export const ChainUiComponents__StatusIcons = {
+  good: FiCheckCircle,
+  bad: FiXCircle,
+  neutral: FiAlertCircle,
+};
